+++ app/Filament/Resources/Teams/Pages/ListTeams.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Filament\Resources\Teams\Pages;
+
+use App\Filament\Concerns\HidesBreadcrumbs;
+use App\Filament\Resources\Teams\TeamResource;
+use Filament\Actions\CreateAction;
+use Filament\Resources\Pages\ListRecords;
+
+class ListTeams extends ListRecords
+{
+    use HidesBreadcrumbs;
+
+    protected static string $resource = TeamResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            CreateAction::make(),
+        ];
+    }
+}